added resume download option
