+++ src/Components/Home/ContactMe/ReachMe/ReachMe.js
@@ -1,6 +1,7 @@
 import React from "react";
 import ReachMeLinks from "./ReachMeLinks/ReachMeLinks";
 import { Row, Col } from "react-bootstrap";
+import ResumeLink from "./ResumeLink/ResumeLink";
 
 export default function ReachMe() {
     return (
@@ -11,7 +12,7 @@ export default function ReachMe() {
             </Col>
             <Col md={6} className="resume-download">
                 {/* Resume download component */}
-                
+                <ResumeLink></ResumeLink>
             </Col>
         </Row>
     )
