+++ src/Components/Home/ContactMe/ReachMe/ResumeLink/ResumeLink.scss
@@ -1,0 +1,18 @@
+.download-button{
+    border: 2px solid var(--info);
+    background-color: var(--dark);
+    border-radius: 0;
+    padding: 10px;
+    padding-left: 40px;
+    padding-right: 40px;
+    font-size: x-large;
+    text-decoration: none;
+    color: var(--info);
+}
+
+.download-button:hover{
+    border: 2px solid var(--light);
+    color: var(--light);
+    background-color: var(--dark);
+    text-decoration: none;
+}
